requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Cors>=4.0
PyMySQL>=1.1
python-dotenv>=1.0
gunicorn>=21.0
cryptography>=42.0
requests>=2.32
